Move comon code from isodoc/*_convert.rb to single place
